flexible/requirements.txt
Flask==0.12.2
pyjwt==1.5.3
flask-cors==3.0.3
google-auth==1.1.1
requests==2.18.4
gunicorn==19.7.1
urlfetch==1.0.2
mysql-python==1.2.5
